requirements.txt
streamlit

python-dotenv

langchain
langchain-community
langchain-text-splitters
langchain-huggingface
langchain-chroma
langchain-groq
langchain-classic

pypdf

chromadb

sentence-transformers
